requirements.txt
streamlit
google-genai
openai
python-dotenv
pillow
pandas
fpdf
regex